fixed codacy
