+++ src/main/java/duke/logic/parsers/commandparser/CommandParser.java
@@ -1,8 +1,14 @@
 package duke.logic.parsers.commandparser;
 
+import duke.commons.Messages;
 import duke.commons.exceptions.DukeException;
+import duke.commons.exceptions.InputNotIntException;
+import duke.commons.exceptions.QueryOutOfBoundsException;
+
+import java.util.logging.Logger;
 
 public abstract class CommandParser<T> {
+    protected static final Logger logger = Logger.getLogger(Logger.GLOBAL_LOGGER_NAME);
 
     public abstract T parse() throws DukeException;
