+++ src/main/java/duke/logic/parsers/commandparser/RouteListParser.java
@@ -3,7 +3,6 @@ package duke.logic.parsers.commandparser;
 import duke.commons.exceptions.DukeException;
 import duke.logic.commands.Command;
 import duke.logic.commands.RouteListCommand;
-import duke.logic.commands.RouteShowCommand;
 import duke.logic.parsers.ParserUtil;
 
 /**
