+++ src/main/java/duke/logic/parsers/commandparser/RouteNodeListParser.java
@@ -3,7 +3,6 @@ package duke.logic.parsers.commandparser;
 import duke.commons.exceptions.DukeException;
 import duke.logic.commands.Command;
 import duke.logic.commands.RouteNodeListCommand;
-import duke.logic.commands.RouteNodeShowCommand;
 import duke.logic.parsers.ParserUtil;
 
 /**
